requirements.txt
accelerate==1.6.0
confection==0.1.5
google-cloud-vision==3.10.1
h11==0.14.0
huggingface-hub==0.30.2
idna==3.10
isodate==0.6.1
Jinja2==3.1.6
joblib==1.4.2
language_data==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
numpy==2.2.5
nvidia-cuda-runtime-cu12==12.6.77
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.21.1
openpyxl==3.1.5
pi_heif==0.22.0
pikepdf==9.7.0
prov==2.0.1
psutil==7.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
spacy==3.8.5
timm==1.0.15
tokenizers==0.21.1
typing_extensions==4.13.2

pandas>=1.0.0
tqdm
torch
numpy
python-dateutil
spacy>=3.0.0,<4.0.0
faiss-cpu
faiss-gpu
transformers
sentence-transformers
tiktoken==0.5.1
datasets
sentencepiece
bitsandbytes>=0.42.0

fastapi
uvicorn[standard]
pydantic

unstructured[all]
unstructured[pdf]
python-multipart

huggingface_hub[hf_xet]